MAINT: Fix miscellaneous sign-compare warnings.
Small fixes scattered over six files.

- numpy/core/src/multiarray/array_assign_scalar.c
- numpy/core/src/multiarray/common.c
- numpy/core/src/multiarray/multiarray_tests.c.src
- numpy/core/src/multiarray/refcount.c
- numpy/core/src/multiarray/shape.c
- numpy/core/src/multiarray/temp_elide.c

+++ numpy/core/src/multiarray/array_assign_scalar.c
@@ -233,7 +233,7 @@ PyArray_AssignRawScalar(PyArrayObject *dst,
          * Use a static buffer to store the aligned/cast version,
          * or allocate some memory if more space is needed.
          */
-        if (sizeof(scalarbuffer) >= PyArray_DESCR(dst)->elsize) {
+        if ((int)sizeof(scalarbuffer) >= PyArray_DESCR(dst)->elsize) {
             tmp_src_data = (char *)&scalarbuffer[0];
         }
         else {
